DataViews Extensibility: Allow unregistering the duplicate pattern action
